SmartRoute-Cloud-Engine/requirements.txt
django==6.0.3
djangorestframework==3.15.1
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
azure-storage-queue